:recycle: Make ttt_create_board consistent by preventing UB
- Refactor rename ttt_initialize_cells to ttt_reset_cells
- Change behavior of ttt_create_board by making it initialize cells value

Those changes has been done to remove useless undefined behavior (UB) that can occured by not calling ttt_initialize_cells (now ttt_reset_cells) before to use it. In fact, is your board is greatly scoped, you don't need to seperate creation of board and initialization of its cells. ttt_reset_cells is not an internal function because it allows us to avoid to destroy and create a new board